+++ database/migrations/2015_09_10_185135_add_reminder_emails.php
@@ -42,22 +42,24 @@ class AddReminderEmails extends Migration
     public function down()
     {
         Schema::table('accounts', function ($table) {
-            $table->dropColumn('email_subject_invoice');
-            $table->dropColumn('email_subject_quote');
-            $table->dropColumn('email_subject_payment');
-            
-            $table->dropColumn('email_subject_reminder1');
-            $table->dropColumn('email_subject_reminder2');
-            $table->dropColumn('email_subject_reminder3');
-            
-            $table->dropColumn('email_template_reminder1');
-            $table->dropColumn('email_template_reminder2');
-            $table->dropColumn('email_template_reminder3');
-            
+            if (Schema::hasColumn('accounts', 'email_subject_invoice')) {
+                $table->dropColumn('email_subject_invoice');
+                $table->dropColumn('email_subject_quote');
+                $table->dropColumn('email_subject_payment');
+
+                $table->dropColumn('email_subject_reminder1');
+                $table->dropColumn('email_subject_reminder2');
+                $table->dropColumn('email_subject_reminder3');
+
+                $table->dropColumn('email_template_reminder1');
+                $table->dropColumn('email_template_reminder2');
+                $table->dropColumn('email_template_reminder3');
+            }
+
             $table->dropColumn('enable_reminder1');
             $table->dropColumn('enable_reminder2');
             $table->dropColumn('enable_reminder3');
-            
+
             $table->dropColumn('num_days_reminder1');
             $table->dropColumn('num_days_reminder2');
             $table->dropColumn('num_days_reminder3');
